refactor(quality): pass-3 cleanups — union hooks live with their store
- useForeignSessionLaneId moves into state/crossMachineLanes.ts so every
  root-store union reader lives beside the data and its invariant
- ChatPrPane uses useMachineEntryForBinding instead of a hand-rolled
  root-store read (machineEntryForBinding now has zero raw React callers)
- drop three unnecessary casts in the AppStoreProvider regression tests
- refresh the LaneGitActionsPane machine-identity comment
